Day 23, Part 2
Answer is correct, but requires a hack, because answer depends on whether the
first node to be removed belongs to the largest complete graph. If it is,
the largest complete graph is not be found.

The hack was to reorder the input so that the first node to be removed does
not belong to the larges complete graph - because without this, the first
node does belong the complete graph.

A systematic way of guaranteeing to find the largest complete graph is to
run the algorithm multiple times, where in each iteration removing a different
node first. We didn't bother doing that, but that is what you would do to
solve the general case.

Another con of the algorith is that if multiple largest complete graphs exist
then only one would be found, and which one would depend on the input order.